requirements.txt

streamlit>=1.37.0
PyYAML>=6.0.1
networkx>=3.1
python-dotenv>=1.0.1
pydantic>=2.8.2
jsonschema>=4.23.0
# openai>=1.40.0   # Uncomment when wiring LLM
